+++ src/TaskrowSharp/Models/ClientModels/ClientAdministrativeDetail.cs
@@ -4,8 +4,8 @@ public class ClientAdministrativeDetail
 {
     public int ClientAdministrativeDetailID { get; set; }
     public string Memo { get; set; }
-    public double? AdComissionPercentage { get; set; }
-    public double? SupplierComissionPercentage { get; set; }
-    public double? BVPercentage { get; set; }
+    public decimal? AdComissionPercentage { get; set; }
+    public decimal? SupplierComissionPercentage { get; set; }
+    public decimal? BVPercentage { get; set; }
     public bool AddComissionTax { get; set; }
 }
